move UI rendering out of App.tsx this makes bundles that just use API components (eg. to render out PUI settings controls for inclusion CUI) much easier
