+++ src/frontend/src/App.tsx
@@ -1,9 +1,6 @@
-import { useViewportSize } from '@mantine/hooks';
 import { QueryClient } from '@tanstack/react-query';
 import axios from 'axios';
-import { lazy } from 'react';
 
-import { Loadable } from './functions/loading';
 import { useLocalState } from './states/LocalState';
 import { useSessionState } from './states/SessionState';
 
@@ -18,23 +15,3 @@ export function setApiDefaults() {
   api.defaults.headers.common['Authorization'] = `Token ${token}`;
 }
 export const queryClient = new QueryClient();
-
-function checkMobile() {
-  const { height, width } = useViewportSize();
-  if (width < 425 || height < 425) return true;
-  return false;
-}
-
-const MobileAppView = Loadable(lazy(() => import('./views/MobileAppView')));
-const DesktopAppView = Loadable(lazy(() => import('./views/DesktopAppView')));
-
-// Main App
-export default function App() {
-  // Check if mobile
-  if (checkMobile()) {
-    return <MobileAppView />;
-  }
-
-  // Main App component
-  return <DesktopAppView />;
-}
