+++ src/frontend/src/main.tsx
@@ -4,8 +4,8 @@ import ReactDOM from 'react-dom/client';
 import 'react-grid-layout/css/styles.css';
 import 'react-resizable/css/styles.css';
 
-import App from './App';
 import { HostList } from './states/states';
+import MainView from './views/MainView';
 
 // define settings
 declare global {
@@ -71,7 +71,7 @@ export const base_url = window.INVENTREE_SETTINGS.base_url || 'platform';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
-    <App />
+    <MainView />
   </React.StrictMode>
 );
